# STORE/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "STORE",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/inventory/ingredients": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "List Ingredients",
        "operationId": "list_ingredients_api_v1_inventory_ingredients_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "lowStock",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lowstock"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Create Ingredient",
        "operationId": "create_ingredient_api_v1_inventory_ingredients_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/ingredients/{ingredient_id}": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Ingredient",
        "operationId": "get_ingredient_api_v1_inventory_ingredients__ingredient_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ingredient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "inventory"
        ],
        "summary": "Update Ingredient",
        "operationId": "update_ingredient_api_v1_inventory_ingredients__ingredient_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ingredient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "inventory"
        ],
        "summary": "Delete Ingredient",
        "operationId": "delete_ingredient_api_v1_inventory_ingredients__ingredient_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ingredient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/low-stock": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Low Stock",
        "operationId": "low_stock_api_v1_inventory_low_stock_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inventory/purchase": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Purchase",
        "operationId": "purchase_api_v1_inventory_purchase_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inventory/adjust": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Adjust",
        "operationId": "adjust_api_v1_inventory_adjust_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inventory/transfer": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Transfer",
        "operationId": "transfer_api_v1_inventory_transfer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inventory/allocate": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Allocate",
        "operationId": "allocate_api_v1_inventory_allocate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inventory/return": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Return Stock",
        "operationId": "return_stock_api_v1_inventory_return_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inventory/history/{ingredient_id}": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "History",
        "operationId": "history_api_v1_inventory_history__ingredient_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ingredient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/allocate-food": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Allocate Food",
        "description": "Allocate N units of a food item to a store.\n\nResolves the food item's active recipe, computes the required ingredients\nper unit \u00d7 N, and atomically deducts from the store's inventory.\nRolls back the entire allocation if any ingredient is insufficient.\nNow writes a persistent record to the ``allocations`` collection.",
        "operationId": "allocate_food_api_v1_inventory_allocate_food_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodAllocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdjustmentRequest": {
        "properties": {
          "ingredientId": {
            "type": "string",
            "title": "Ingredientid"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          },
          "storeId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storeid"
          }
        },
        "type": "object",
        "required": [
          "ingredientId",
          "quantity",
          "reason"
        ],
        "title": "AdjustmentRequest"
      },
...
      "AllocationRequest": {
        "properties": {
          "ingredientId": {
            "type": "string",
            "title": "Ingredientid"
          },
          "storeId": {
            "type": "string",
            "title": "Storeid"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "date": {
            "type": "string",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "ingredientId",
          "storeId",
          "quantity",
          "date"
        ],
        "title": "AllocationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngredientCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "unit": {
            "type": "string",
            "minLength": 1,
            "title": "Unit"
          },
          "currentStock": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentstock"
          },
          "minimumStock": {
            "type": "number",
            "title": "Minimumstock",
            "default": 0
          },
          "maximumStock": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximumstock"
          },
          "supplierId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplierid"
          },
          "purchasePrice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchaseprice"
          },
          "costPerUnit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Costperunit"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          }
        },
        "type": "object",
        "required": [
          "name",
          "unit"
        ],
        "title": "IngredientCreateRequest"
      },
      "IngredientUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "currentStock": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentstock"
          },
          "minimumStock": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimumstock"
          },
          "maximumStock": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximumstock"
          },
          "supplierId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplierid"
          },
          "purchasePrice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchaseprice"
          },
          "costPerUnit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Costperunit"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          }
        },
        "type": "object",
        "title": "IngredientUpdateRequest"
      },
...
      "PurchaseRequest": {
        "properties": {
          "ingredientId": {
            "type": "string",
            "title": "Ingredientid"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "purchasePrice": {
            "type": "number",
            "title": "Purchaseprice"
          },
          "supplierId": {
            "type": "string",
            "title": "Supplierid"
          }
        },
        "type": "object",
        "required": [
          "ingredientId",
          "quantity",
          "unit",
          "purchasePrice",
          "supplierId"
        ],
        "title": "PurchaseRequest"
      },
...
      "ReturnRequest": {
        "properties": {
          "ingredientId": {
            "type": "string",
            "title": "Ingredientid"
          },
          "storeId": {
            "type": "string",
            "title": "Storeid"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "date": {
            "type": "string",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "ingredientId",
          "storeId",
          "quantity",
          "date"
        ],
        "title": "ReturnRequest"
      },
...
      "TransferRequest": {
        "properties": {
          "ingredientId": {
            "type": "string",
            "title": "Ingredientid"
          },
          "fromStoreId": {
            "type": "string",
            "title": "Fromstoreid"
          },
          "toStoreId": {
            "type": "string",
            "title": "Tostoreid"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "ingredientId",
          "fromStoreId",
          "toStoreId",
          "quantity"
        ],
        "title": "TransferRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}